pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rss-watcher"
version = "0.1.0"
description = "Self-hosted Linux kernel and Android update watcher"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "apscheduler>=3.10,<4.0",
  "beautifulsoup4>=4.12,<5.0",
  "fastapi>=0.116,<1.0",
  "httpx>=0.28,<1.0",
  "psycopg[binary]>=3.2,<4.0; python_version < '3.14'",
  "pydantic-settings>=2.8,<3.0",
  "pyyaml>=6.0.2,<7.0",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn>=0.35,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
